netconf: Reorder object and interface name in ethtool(8) params string
Our generic format for references is [<new_name>|<object>]@<old_name>
while for ethtool(8) we used <new_name>-<object> that does not fit
in references concept.

Reorder fields in params string to match concept.

Before:
-------

  # apply ethtool settings
  en2_ref1="en2-K gro off gso off"
  en2_ref2="en2--set-ring rx 1024 tx 1024"

After:
------

  # apply ethtool settings
  en2_ref1="-K-en2 gro off gso off"
  en2_ref2="--set-ring-en2 rx 1024 tx 1024"

So the only sysctl(8) params are now out of references format concept but
this is less important as sysctl params string passed to sysctl(8) as is.

Fixes: acbbb6d793c4 ("netconf: Move ethtool(8) settings to references")
Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>