netconf: Label maximum possible group number as invalid
It is now marked with _reserved label to indicate that it should not
be used somewhere in netconf or netctl, label it with invalid to
indicate that it must not be used to configure network device group
neither via netconf nor via other means.

This label is used by rebird project as special value to indicate
that variable isn't assigned any value.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>